a few more changes regarding 32bit vs 64bit values

+++ daemon/util/Util.cpp
@@ -886,14 +886,14 @@ void Util::Sleep(int ms)
 	std::this_thread::sleep_for(std::chrono::milliseconds(ms));
 }
 
-int64 WebUtil::DecodeBase64(char* inputBuffer, int64 inputBufferLength, char* outputBuffer)
+int32 WebUtil::DecodeBase64(char* inputBuffer, int32 inputBufferLength, char* outputBuffer)
 {
-	int64 InputBufferIndex  = 0;
-	int64 OutputBufferIndex = 0;
-	int64 InputBufferLength = inputBufferLength > 0 ? inputBufferLength : strlen(inputBuffer);
+	int32 InputBufferIndex  = 0;
+	int32 OutputBufferIndex = 0;
+	int32 InputBufferLength = inputBufferLength > 0 ? inputBufferLength : strlen(inputBuffer);
 
 	char ByteQuartet [4];
-	int64 i = 0;
+	int32 i = 0;
 	while (InputBufferIndex < InputBufferLength)
 	{
 		// Ignore all characters except the ones in BASE64_ALPHABET
@@ -1107,7 +1107,7 @@ BreakLoop:
 	*output = '\0';
 }
 
-const char* WebUtil::XmlFindTag(const char* xml, const char* tag, int64* valueLength)
+const char* WebUtil::XmlFindTag(const char* xml, const char* tag, int32* valueLength)
 {
 	BString<100> openTag("<%s>", tag);
 	BString<100> closeTag("</%s>", tag);
@@ -1126,21 +1126,21 @@ const char* WebUtil::XmlFindTag(const char* xml, const char* tag, int64* valueLe
 	const char* pend = strstr(pstart, closeTag);
 	if (!pend) return nullptr;
 
-	int64 tagLen = strlen(openTag);
+	int32 tagLen = strlen(openTag);
 	*valueLength = (int64)(pend - pstart - tagLen);
 
 	return pstart + tagLen;
 }
 
-bool WebUtil::XmlParseTagValue(const char* xml, const char* tag, char* valueBuf, int64 valueBufSize, const char** tagEnd)
+bool WebUtil::XmlParseTagValue(const char* xml, const char* tag, char* valueBuf, int32 valueBufSize, const char** tagEnd)
 {
-	int64 valueLen = 0;
+	int32 valueLen = 0;
 	const char* value = XmlFindTag(xml, tag, &valueLen);
 	if (!value)
 	{
 		return false;
 	}
-	int64 len = valueLen < valueBufSize ? valueLen : valueBufSize - 1;
+	int32 len = valueLen < valueBufSize ? valueLen : valueBufSize - 1;
 	strncpy(valueBuf, value, len);
 	valueBuf[len] = '\0';
 	if (tagEnd)
@@ -1384,7 +1384,7 @@ BreakLoop:
 	*output = '\0';
 }
 
-const char* WebUtil::JsonFindField(const char* jsonText, const char* fieldName, int64* valueLength)
+const char* WebUtil::JsonFindField(const char* jsonText, const char* fieldName, int32* valueLength)
 {
 	BString<100> openTag("\"%s\"", fieldName);
 
@@ -1396,7 +1396,7 @@ const char* WebUtil::JsonFindField(const char* jsonText, const char* fieldName, 
 	return JsonNextValue(pstart, valueLength);
 }
 
-const char* WebUtil::JsonNextValue(const char* jsonText, int64* valueLength)
+const char* WebUtil::JsonNextValue(const char* jsonText, int32* valueLength)
 {
 	const char* pstart = jsonText;
 
@@ -1430,7 +1430,7 @@ const char* WebUtil::JsonNextValue(const char* jsonText, int64* valueLength)
 		ch = *++pend;
 	}
 
-	*valueLength = (int64)(pend - pstart);
+	*valueLength = (int32)(pend - pstart);
 	return pstart;
 }
